+++ src/main/java/org/folio/linked/data/configuration/json/deserialization/instance/IdentifierFieldDeserializer.java
@@ -4,7 +4,6 @@ import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_IAN;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_ISBN;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_ISSN;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_LCCN;
-import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_LOCAL;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_UNKNOWN;
 
 import com.fasterxml.jackson.core.JsonParser;
@@ -17,7 +16,6 @@ import org.folio.linked.data.domain.dto.IdentifierField;
 import org.folio.linked.data.domain.dto.IsbnField;
 import org.folio.linked.data.domain.dto.IssnField;
 import org.folio.linked.data.domain.dto.LccnField;
-import org.folio.linked.data.domain.dto.LocalIdField;
 import org.folio.linked.data.domain.dto.OtherIdField;
 import org.folio.linked.data.exception.RequestProcessingExceptionBuilder;
 import org.folio.linked.data.util.DtoDeserializer;
@@ -29,7 +27,6 @@ public class IdentifierFieldDeserializer extends JsonDeserializer<IdentifierFiel
     ID_ISSN.getUri(), IssnField.class,
     ID_ISBN.getUri(), IsbnField.class,
     ID_IAN.getUri(), IanField.class,
-    ID_LOCAL.getUri(), LocalIdField.class,
     ID_UNKNOWN.getUri(), OtherIdField.class
   );
   private final DtoDeserializer<IdentifierField> dtoDeserializer;